+++ bc-cache.php
@@ -9,7 +9,7 @@
  * Author URI: https://www.chesio.com
  * Requires PHP: 8.0
  * Requires WP: 6.1
- * Tested up to: 6.2
+ * Tested up to: 6.3
  * Text Domain: bc-cache
  * GitHub Plugin URI: https://github.com/chesio/bc-cache
  * Update URI: https://github.com/chesio/bc-cache